requirements.txt
Django==2.2.8
django-crispy-forms==1.8.1
django-tinymce4-lite==1.7.5
jsmin==2.2.2
mysqlclient==1.4.6
Pillow==7.0.0
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.0
